Drop log level of CRL download failures to info
Once this root cert expires I expect the CRLs will be taken offline, after that time the CRL becomes useless.

In a few years we don't want every build to suddenly have errors in the logs.

I think this is acceptable as your web browser or OS will not inform you of CRL download failures.